Extracted functionality to own files
