+++ .vscode/settings.json
@@ -1,3 +1,10 @@
 {
-    "prettier.tabWidth": 4
+    "prettier.tabWidth": 4,
+    "cSpell.enableFiletypes": [
+        "swift"
+    ],
+    "cSpell.words": [
+        "qbio",
+        "struct"
+    ]
 }
