+++ Sources/qbio-lib/AudioDB/Artist.swift
@@ -1,0 +1,7 @@
+struct Artist: Codable {
+    let strBiographyEN: String
+}
+
+struct ArtistQueryResponse: Codable {
+    let artists: [Artist]?
+}
